[fix]: Fix float values in kinematic links

+++ src/KinematicLink/DegStepLink.h
@@ -5,16 +5,17 @@
 class DegStepLink : public BaseLink {
 private:
     uint16_t steps_per_rev;
-    uint16_t deg_per_step;
+    float deg_per_step;
 
 public:
     explicit DegStepLink(const uint16_t steps_per_rev) : BaseLink(), steps_per_rev(steps_per_rev),
-                                                         deg_per_step(360 / steps_per_rev) {
+                                                         deg_per_step(360.0f / static_cast<float>(steps_per_rev)) {
     }
 
     DegStepLink(const uint16_t steps_per_rev, BaseLink *previos_link) : BaseLink(previos_link),
                                                                         steps_per_rev(steps_per_rev),
-                                                                        deg_per_step(360 / steps_per_rev) {
+                                                                        deg_per_step(360.0f /
+                                                                            static_cast<float>(steps_per_rev)) {
     }
 
     /**
